+++ CHANGELOG.rst
@@ -50,6 +50,7 @@ Added
     - ``CurrentUserFormMixin`` for providing current using to forms as ``current_user``
     - Helper mixin ``KnoxAuthMixin`` for views testing
     - Sanitize search input (#332)
+    - Handle project list title cell overflow (#306)
 - **Samplesheets**
     - Display original study/assay filenames as tooltips (#283)
     - Display assays for samples in search results (#157)